move GO data to separate props
